[RecoPixelVertexing] Move public headers to interface directory
